+++ android/app/src/test/java/poke/rogue/helper/remote/dto/response/pokemon/PokemonResponseTest.kt
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.Test
 import poke.rogue.helper.data.model.Pokemon
 import poke.rogue.helper.data.model.Type
 import poke.rogue.helper.data.model.toData
-import poke.rogue.helper.pokemonResponses
-import poke.rogue.helper.pokemonTypeResponses
+import poke.rogue.helper.utils.fixture.pokemonResponses
+import poke.rogue.helper.utils.fixture.pokemonTypeResponses
 
 class PokemonResponseTest {
     @Test
